+++ tests/testthat/test-check_experiments.R
@@ -1,8 +1,6 @@
-context("test-check_experiments")
-
 test_that("use", {
 
-  expect_silent(check_experiments(experiments = list(create_experiment())))
+  expect_silent(check_experiments(list(create_experiment())))
 })
 
 test_that("each element in the list must be a proper experiment", {
@@ -33,10 +31,9 @@ test_that("correct order of experiments", {
   if (rappdirs::app_dir()$os == "win") return()
 
   gen_exp <- create_test_gen_experiment()
-  check_experiment(gen_exp)
-
   cand_exp <- create_test_cand_experiment()
-  check_experiment(cand_exp)
+  # Must use different inference model than generative model
+  cand_exp$inference_model$site_model <- beautier::create_gtr_site_model()
 
   # OK order
   expect_silent(
@@ -69,17 +66,17 @@ test_that("same beast2_options_filenames and error fileanames in candidates", {
   cand_experiment_1 <- create_test_cand_experiment()
   cand_experiment_2 <- create_test_cand_experiment()
   gen_experiment <- create_test_gen_experiment()
+  # Must use different inference model than generative model
+  cand_experiment_1$inference_model$site_model <-
+    beautier::create_tn93_site_model()
+  cand_experiment_2$inference_model$site_model <-
+    beautier::create_gtr_site_model()
+
   expect_silent(
     check_experiments(list(gen_experiment, cand_experiment_1))
   )
   expect_silent(
     check_experiments(list(gen_experiment, cand_experiment_2))
-  )
-  expect_silent(
-    check_experiments(list(cand_experiment_1, cand_experiment_1))
-  )
-  expect_silent(
-    check_experiments(list(cand_experiment_2, cand_experiment_2))
   )
   expect_error(
     check_experiments(list(cand_experiment_1, cand_experiment_2)),
@@ -117,7 +114,12 @@ test_that("differ each beast2_options_filename and error filename", {
   if (rappdirs::app_dir()$os == "win") return()
 
   experiment_1 <- create_test_cand_experiment()
-  experiment_2 <- experiment_1
+  good_experiment_2 <- experiment_1
+  # Experiments must have different inference models
+  good_experiment_2$inference_model$site_model <-
+    beautier::create_gtr_site_model()
+
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
 
   # BEAST2 input filename
@@ -126,7 +128,7 @@ test_that("differ each beast2_options_filename and error filename", {
     check_experiments(list(experiment_1, experiment_2)),
     "Candidate models must have same BEAST2 input filename"
   )
-  experiment_2 <- experiment_1
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
 
   # BEAST2 output log filename
@@ -135,7 +137,7 @@ test_that("differ each beast2_options_filename and error filename", {
     check_experiments(list(experiment_1, experiment_2)),
     "Candidate models must have same BEAST2 output log filename"
   )
-  experiment_2 <- experiment_1
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
 
   # BEAST2 output trees filenames
@@ -144,7 +146,7 @@ test_that("differ each beast2_options_filename and error filename", {
     check_experiments(list(experiment_1, experiment_2)),
     "Candidate models must have same BEAST2 output trees filename"
   )
-  experiment_2 <- experiment_1
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
 
   # BEAST2 output state
@@ -153,7 +155,7 @@ test_that("differ each beast2_options_filename and error filename", {
     check_experiments(list(experiment_1, experiment_2)),
     "Candidate models must have same BEAST2 output state filename"
   )
-  experiment_2 <- experiment_1
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
 
   # Errors filename
@@ -162,6 +164,42 @@ test_that("differ each beast2_options_filename and error filename", {
     check_experiments(list(experiment_1, experiment_2)),
     "Candidate models must have same errors filename"
   )
-  experiment_2 <- experiment_1
+  experiment_2 <- good_experiment_2
   expect_silent(check_experiments(list(experiment_1, experiment_2)))
+})
+
+test_that("detect same model in generative and candidate model", {
+
+  if (rappdirs::app_dir()$os == "win") return()
+
+  inference_model <- create_inference_model()
+  gen_exp <- create_test_gen_experiment(
+    inference_model = inference_model
+  )
+  check_experiment(gen_exp)
+
+  cand_exp <- create_test_cand_experiment(
+    inference_model = inference_model
+  )
+  check_experiment(cand_exp)
+
+  experiments <- list(gen_exp, cand_exp)
+
+  expect_error(
+    check_experiments(experiments),
+    "All inference models must be unique"
+  )
+})
+
+test_that("detect same model in two candidate models", {
+
+  if (rappdirs::app_dir()$os == "win") return()
+
+  cand_exp_1 <- create_test_cand_experiment()
+  cand_exp_2 <- cand_exp_1
+  experiments <- list(cand_exp_1, cand_exp_2)
+  expect_error(
+    check_experiments(experiments),
+    "All inference models must be unique"
+  )
 })