+++ tests/testthat/test-create_pir_params_twin.R
@@ -26,7 +26,12 @@ test_that("two candidates", {
   )
   experiments[[1]]$beast2_options <- experiments[[2]]$beast2_options
   experiments[[1]]$errors_filename <- experiments[[2]]$errors_filename
-  check_experiments(experiments)
+  # Experiments must have different inference models
+  experiments[[1]]$inference_model$site_model <-
+    beautier::create_tn93_site_model()
+  experiments[[2]]$inference_model$site_model <-
+    beautier::create_gtr_site_model()
+
   pir_params <- create_test_pir_params(
     experiments = experiments,
     twinning_params = create_twinning_params()
@@ -42,9 +47,14 @@ test_that("one generative, two candidates", {
     create_test_cand_experiment(),
     create_test_cand_experiment()
   )
+  # Candidates must have different model than generative model
+  experiments[[2]]$inference_model$site_model <-
+    beautier::create_tn93_site_model()
+  experiments[[3]]$inference_model$site_model <-
+    beautier::create_gtr_site_model()
+  # Candidates must share same BEAST2 options
   experiments[[2]]$beast2_options <- experiments[[3]]$beast2_options
   experiments[[2]]$errors_filename <- experiments[[3]]$errors_filename
-  check_experiments(experiments)
 
   pir_params <- create_test_pir_params(
     experiments = experiments,