+++ NEWS.md
@@ -2,6 +2,61 @@
 
 Newest versions at top.
 
+## `pirouette` 1.5 (2019-08-19)
+
+### NEW FEATURES
+
+ * None
+
+### MINOR IMPROVEMENTS
+
+ * Checks that all experiments have a unique inference model
+
+### BUG FIXES
+
+ * None
+
+### DEPRECATED AND DEFUNCT
+
+ * None
+
+## `pirouette` 1.4.1 (2019-08-19)
+
+### NEW FEATURES
+
+ * None
+
+### MINOR IMPROVEMENTS
+
+ * None
+
+### BUG FIXES
+
+ * Table with evidences for twin models works
+
+### DEPRECATED AND DEFUNCT
+
+ * None
+
+## `pirouette` 1.4 (2019-08-18)
+
+### NEW FEATURES
+
+ * Evidence estimations shows estimated error of evidence
+ * Evidence estimations shows effective sample size in estimating the evidence
+
+### MINOR IMPROVEMENTS
+
+ * Better error message when using a CBS site model and too few taxa
+
+### BUG FIXES
+
+ * None
+
+### DEPRECATED AND DEFUNCT
+
+ * None
+
 ## `pirouette` 1.3 (2019-08-15)
 
 ### NEW FEATURES